pyproject.toml
[project]
name = "reddit-contextual-agent"
version = "0.2.0"
description = "Reddit scraper and data pipeline for Contextual AI Agent Composer"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "praw>=7.7.0",
    "contextual-client>=0.6.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "structlog>=24.0.0",
    "httpx>=0.27.0",
    "asyncpg>=0.29.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[project.scripts]
reddit-agent = "reddit_agent.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/reddit_agent"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
